separate cli fixtures from the tests
